Move schemas into topline module namespace
